support form parameters with the new enumerable system
