+++ DragonFruit.Common.Data/Utils/ParameterUtils.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. Please refer to the LICENSE file at the root of this project for details
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
@@ -22,18 +23,68 @@ namespace DragonFruit.Common.Data.Utils
         /// </summary>
         internal static IEnumerable<KeyValuePair<string, string>> GetParameter<T>(object host, CultureInfo culture) where T : IProperty
         {
-            var type = typeof(T);
-
-            foreach (var property in host.GetType().GetProperties(DefaultFlags))
+            foreach (var property in host.GetType().GetProperties(ParameterUtils.DefaultFlags))
             {
-                if (!(Attribute.GetCustomAttribute(property, type) is T parameter))
+                if (!(Attribute.GetCustomAttribute(property, typeof(T)) is T parameter))
                 {
                     continue;
                 }
 
-                var convertedValue = property.GetValue(host).AsString(culture);
-                if (convertedValue != null)
+                // check if the type we've got is an IEnumerable of anything (in this case object)
+                if (typeof(IEnumerable).IsAssignableFrom(property.PropertyType))
+                {
+                    var values = (IEnumerable)property.GetValue(host);
+
+                    switch (parameter.CollectionHandling)
+                    {
+                        case CollectionConversionMode.Recursive:
+                        {
+                            foreach (var item in values)
+                            {
+                                // return multiple of the same key
+                                yield return new KeyValuePair<string, string>(parameter.Name, item.AsString(culture));
+                            }
+
+                            break;
+                        }
+
+                        case CollectionConversionMode.Unordered:
+                        {
+                            foreach (var item in values)
+                            {
+                                // return multiple suffixed keys
+                                yield return new KeyValuePair<string, string>($"{parameter.Name}[]", item.AsString(culture));
+                            }
+
+                            break;
+                        }
+
+                        // if it's order-explicit we need to use an enumerator as there's no length count
+                        case CollectionConversionMode.Ordered:
+                        {
+                            var counter = 0;
+                            var enumerator = values.GetEnumerator();
+
+                            while (enumerator.MoveNext())
+                            {
+                                // return suffixed version with counter
+                                yield return new KeyValuePair<string, string>($"{parameter.Name}[{counter}]", enumerator.Current.AsString(culture));
+                                counter++;
+                            }
+
+                            break;
+                        }
+
+                        default:
+                            throw new ArgumentOutOfRangeException();
+                    }
+                }
+                else
+                {
+                    // anything else can just be taken as a single object for now
+                    var convertedValue = property.GetValue(host).AsString(culture);
                     yield return new KeyValuePair<string, string>(parameter.Name, convertedValue);
+                }
             }
         }
